Relax permission checks in nnrpd for TLS private key
Second-guessing whether we can read the file can get us in trouble
if file system ACLs are in use, and requiring the group match the
group of nnrpd makes it harder than necessary to set up automatic
certificate renewal via something like Let's Encrypt.  Relax the
permission checks to just check that the private key isn't
world-readable and that we can read it.