Add support for a new syntaxchecks parameter in inn.conf
This parameter permits to control the level of checks performed by
innd and nnrpd.  Up to now, only one check can be enabled/disabled:
when "laxmid" is mentioned in the values of this new parameter, INN
accepts Message-IDs that contain ".." in the left part, as well as
Message-IDs with two "@" (such Message-IDs would otherwise be considered
as syntactically invalid).