mailpost: add new "-z" parameter to remove header fields
Adds a "-z" parameter to remove any undesired headers from the gated
message.

Also, edits the incoming SMTP message's "Received:" header to remove any "for
<mailbox>" clause which contains the name of the news server (or its
domain).  It determines the name by examining "pathhost", "fromhost" and
"domain" (from file "inn.conf").  The reason for this is to protect the
gateway mailbox from spammers who may seek to abuse the gateway.  This will
also remove any "*-To:" headers that cite the news server host or domain
name (e.g. "Errors-To:" which are set to report back which user bounced by
using a mailbox pattern in the envelope).  As the gateway mailbox name is
now hidden, spammers will have to discover it via other means.  It's not
meant as an anti-spam measure by itself, but spammers can't spam that which
they can't find.  Since the administrator knows the message was gated, he
knows it was delivered to the gateway mailbox, so information deleted from
the trace header isn't "lost"; just merely hidden.  "For" clauses from
other "Received:" headers not specifying the mail server's host/domain are
left intact.

Thanks to Dieter Stussy for the patch.