Avoid use of SA_LEN in INN code
As part of the original IPv6 work, we adopted the SA_LEN macro to
find the length of a struct sockaddr holding either an IPv4 or an
IPv6 address.  This macro is non-standard, however, and is now
setting off warnings in current versions of GCC.

It doesn't appear to have ever caught on, and we only used it in
three places, two of which were easy to refactor to not use it.
The third requires an if statement because it doesn't pass back the
address length from another function, but it's still simpler than
carrying the SA_LEN macro code.

This prepares INN for a later import of rra-c-util that will drop
the SA_LEN macro and Autoconf probe.