Adjust buffindexed header page size to platform
Previously, if the platform page size was larger than 16KB,
buffindexed would just refuse to work on that platform.  Instead,
increase the header page size to match the platform page size on
those platforms, but still require that the header page size be
a multiple of the platform page size.

Use sysconf(_SC_PAGESIZE) to get the page size instead of
getpagesize(), since it's the POSIX-standard way of doing this.