Use a #pragma to ignore yyunput in flex parser
The declaration of yyunput() with UNUSED was removed when fixing
the source to build with -Wredundant-decls, but it looks like it's
still sometimes needed.  Ignore the unused yyunput function with a
pragma to disable the warning instead, since it's very unlikely we
care about unused functions in the flex input file.