Handle hostnames not in DNS in GetFQDN
If the local hostname isn't in DNS at all, GetFQDN currently returns
NULL.  If a domain was provided but the hostname isn't in DNS, append
the domain to the hostname anyway and return the results.  This is 
possibly still a local configuration error, but it seems like 
friendlier behavior, and also lets the test suite run on hosts with
an unresolvable local hostname.