Stop using com_err directly
com_err is an obsolete way of reporting Kerberos errors.  Add some
additional Kerberos portability code from rra-c-util to allow using
krb5_get_error_message, which is the modern approach, and remove the
additional probing for com_err.h.

Also, if --with-krb5 (or the equivalent for other probes) is given and a
krb5-config was found at the install root, skip the PATH search for
krb5-config to avoid confusing output.  This results in redoing the
probe and using cached output when testing multiple krb5-config
features, but that seems better than the confusing output.

Patch from Russ Allbery.