Update to latest rra-c-util upstream version
The Autoconf m4 macros now check that the library's key header
can be found as well as the library. It removes the need of
several checks added in configure.ac to achieve that.

Also fix INN_LIB_KRB5_OPTIONAL probing in the --enable-reduced-depends case
when linking with libkrb5 is not possible but libcom_err is present.
The probes for com_err were previously not skipped, resulting in a
non-empty KRB5_LIBS containing only -lcom_err, which incorrectly
caused the macro to indicate that Kerberos libraries were present.