Make IOV_MAX Autoconf check more cross-compile-friendly
The IOV_MAX Autoconf check always compiled and ran a program
and relied on the output from that program to indicate that
IOV_MAX was set in limits.h, which causes problems with
cross-compilation.  Instead, use AC_CHECK_DECL to check for the
common case that IOV_MAX is defined in limits.h, and only
compile and run a program for the uncommon case that this fails.

Tested only on Linux, but by changing the probe to look for
something other than IOV_MAX and then for something other than
UIO_MAXIOV, to make sure all the branches work.

Thanks, Helmut Grohne.