Fix syntax in Makefile
GNU make 4.3 emits the following warning:  "ignoring prerequisites on
suffix rule definition".

In previous versions, no warning was emitted and a suffix rule was
created, however all prerequisites were ignored and were not part of
the suffix rule.  Starting with GNU make 4.3, the behaviour is the same,
and in addition a warning is generated.  This commit fixes the syntax.