Update to latest rra-c-util upstream version
Specific fixes for INN builds on Alpine and BSD:
* In the getnameinfo replacement, handle musl libc's gethostbyaddr,
which returns the string conversion of the IP address if the host
doesn't resolve.  This only affects the test suite, since musl libc
supports IPv6 and thus doesn't need this replacement.

* NetBSD prefers reallocarr to reallocarray and only prototypes the
latter if _OPENBSD_SOURCE is defined, which is not one of the macros
Autoconf defines by default.  Work around this by checking whether
reallocarray is declared, not only whether it's available, and
protyping it if it's not declared.  This uses the semi-hidden
deprecated symbol on NetBSD.