Use a fixed buffer size for QIO
The buffer size for QIO is also a limit on the length of an
overview line, and the previous complex approach resulted in a
buffer size of 8KB on nearly all systems, which was too small in
practice.

Stop attempting to be clever and size according to the file 
system block size, which is generally small, and instead use a
fixed and predictable buffer size of 32KB, which shouldn't be a
problem on any modern system.