Add detail for rejected header
When nnrpd rejects a posting because of an invalid header that
isn't part of the header table, include the name of the header
as an aid in debugging, or the full header if no colon could be
found or if the line starts with a colon.

Reported by Adam Sjøgren and Lars Ingebrigtsen.