Improve systemd integration
Revert r10475 (when sd_notify is not defined, its return value
is an integer, not void).

Check the return value and warn if the notification message was not
correctly received (it is a useful information for the administrator,
though sd_notify documentation recommends to ignore the return value
of this call).

Also, use a boolean for the first parameter, which is clearer than
a 0 or 1 int.