ovsqlite: fix build with SQLite 3.8.2
sqlite3_malloc64() was defined only in SQLite 3.8.7.  Switch its calls
to sqlite3_malloc().  As they allocate an error message buffer, it is
safe to assume that the size won't exceed INT_MAX.

Also silent a gcc warning with a "return 0" at the end of the main
function of ovsqlite-server.

Thanks to Bo Lindbergh for the patch.