m4/python.m4: fix the detection of old Python 2.x versions
The current code for probing the presence of Python only workq with
Python 2.7+ and 3.1+ because the "major" attribute to sys.version_info
was only added in these versions.

Also, the construct 'x for y in z' is too recent and not known by
Python 2.3.0, the minimum version required for INN, so I changed it
to a classic map().  If two arguments were not given to the m4 macro,
the script also failed.  Now fixed.