inn.conf:  Add new parameter for listen backlog
Allow to tune the length of the queue of pending connections to innd,
nnrpd and some overview storage methods (ovdb for INN 2.6.5 and both
ovdb and ovsqlite for INN 2.7.0).  The maxlisten parameter now permits
configuring their listen backlog, whose previously hard-coded values
were 128 for nnrpd and 25 for the others, which was not high enough
for some uses.

The default value is now 128 for all of them, and configurable in
inn.conf.

Thanks to Kevin Bowling for the patch.