inn.conf:  Add new parameter for listen backlog
Allow to tune the length of the queue of pending connections to innd,
nnrpd and the ovdb overview storage method.  The maxlisten parameter now
permits configuring their listen backlog, whose previously hard-coded
values were 128 for nnrpd and 25 for the others, which was not high
enough for some uses.

The default value is now 128 for all of them, and configurable in
inn.conf.

Thanks to Kevin Bowling for the patch.