Add support for a GIT snapshot type
On branches, use git describe to construct the version number.
On the main branch, that doesn't do anything useful, so instead
extract the date of the last commit, which hopefully will be good
enough.  These generate stable snapshot names if nothing has
changed in the repository, so that we can publish snapshots only
if nothing has changed.

Keep the old CURRENT and STABLE snapshot types so that we can keep
generating daily snapshots as well if we want to.