Fix double free in tdx-util
The data structure is designed to be cached with a reference count,
and therefore is freed when a search is closed if it hasn't been
added to a cache.  tdx-util was unaware of that and was freeing it
again explicitly, causing a double free and a test failure (but no
security implications, since it happened immediately before process
exit).  Remove the duplicate free.

Add some cleanup to an early return path in the same function.