ovdb:  Improve log message if <pathrun> does not exist
When <pathrun> does not exist, ovdb fails starting with the error:
ovdb_init: OVDB: can not open database unless ovdb_monitor is running

As reported by John Goerzen in Debian bug 993211, it is difficult to
find the root cause.
Now ovdb_init will report the path of the lock file it does not manage
to create.