Build concurrency fixes
We serialize build of top-level subdirectories. The previous attempt
to support make -j was broken, since the subdirectories have dependencies
between them and attempt to cross-invoke one another, leading to multiple
concurrent make invocations on the same directory.

A couple of missing dependencies are also added. The sql-init.h
dependency should be generated by makedepend, but it's confused about
object code extensions. This change is a stopgap until that is
addressed.