Remove alloca from nnrpd
alloca was added in 2004 in the TLS code, but removed in 2008 to use
realloc() because of segmentation faults.
This commit really removes its use from nnrpd in the 2.6 branch.

The whole alloca machinery will be removed in the 2.7 branch from all
source code (commit 5f506335d948ea400cb64ea49182209ab4a32a70).