controlchan: Remove sendsys, senduuname and version support
These control articles are deprecated and should no longer be handled.
Remove the related controlchan modules.

Also remove the "doifarg" keyword from controlchan and control.ctl
documentation (it is only used for these 3 kinds of control messages) but
keep the recognition of the sendsys, senduuname and version control messages
in inncheck and scanlogs (because they are still referenced in control.ctl)
as well as filter_innd.py in the example of a regexp for obsolete control
articles.