Add support for LibreSSL
A few systems like macOS and OpenBSD use LibreSSL by default.
Let's support it and integrate LibreSSL with INN.

We now handle properly OPENSSL_VERSION_NUMBER (which LibreSSL defines
as 2.0.0 -- a major OpenSSL version which never existed).  Some parts
of existing code was not appropriate for a build with LibreSSL.

And of course we also handle LIBRESSL_VERSION_NUMBER.

close #217