Fix (unlikely) possible double free in Cancel-Lock
After freeing the canlock buffer, set it to NULL.  Otherwise, it might be
double freed in an unlikely case that a user posts 3 articles in the same
NNTP session, and libcanlock failed to generate lock hashes for the 2nd
article (but managed for the 1st).  When trying to send the 3rd article,
the canlock buffer will be freed, though unallocated.