No longer generate ChangeLog for releases
We already ship a meaningful NEWS file with releases.  It is no longer useful
to go on generating a full ChangeLog file containing all commit logs.  (Almost
nobody reads it, and it is a waste of space in packages.)

Now that we have migrated to GitHub, it is very easy to browse source code
and, if needed, obtain a similar ChangeLog.  One can even look at code changes
at the same time!

Tag comparison gives that:
  https://github.com/InterNetNews/inn/compare/2.6.3...2.6.4