Reformat Perl code with perltidy
Skip reformatting a few lines in perl/INN/Config.pm.in
and innshellvars.pl.in

Manually fix a few outdented lines because of too long comments or strings.

Use '=>' instead of ',' in the definitions of hashes so that they are
automatically aligned by perltidy.

Manually remove remaining tabulations in comments.