Document news gatewaying to mail in mailpost
How to set it up was not clear at all.  Now document how to set up
approvals (-a flag and A access in readers.conf) and moderation
(in moderators).

Use a better name for the persistent database directory in the
example (/var/tmp was not a good choice).

Mention in -b and -t that mailpost is as often as not run as the
mail user.

Mention which inn.conf parameters are used for the domain name.

Mention that the -a argument is usually an e-mail address.