black: Remove the experimental flag
The Black project has just released their first non-beta release!
Starting with Black 22.1.0, a new stability policy applies.

The --experimental-string-processing no longer exists, so we just
remove it.  It is no longer necessary (though experimental style could now
be enabled with --preview, it will be included in the first release of the
next year, if kept).