Cancel-Lock: new addcanlockuser parameter in readers.conf
Provide a way to deactivate the generation of user-specific hashes
in access groups shared between several people with the same identity.

Update samples and a few examples to highlight that.

Also, reword documentation and comments in source code to clarify the notion
of user identity (a user is not always authenticated, via AUTHINFO, but has
an identity assigned by an auth group).

see #47