incoming.conf: Improve the name of a boolean variables
Rename "nolist", "noresendid" and (still not yet in a release) "noxbatch"
parameters to "list", "resendid" and "xbatch".  These are better names than
with a negation.

At the same time, remove unused "comment" and "email" parameters (they
were just documented as reserved for future use).

innupgrade will take care of the changes (inverting the values of booleans,
commenting lines of removed parameters).

Clarify the behaviour of "resendid" in documentation (the previous wording
was not pretty clear).