Fix the visibility of inet_aton and inet_ntoa replacements
Their visibility was wrongly set to hidden, which would break the build
on systems needing them (very rare nowadays).

According to Russ, inet_ntoa used to be required on IRIX systems because of
an incompatibility between the system ABI and GCC in how structs were passed
on the stack; inet_aton was required for the cygwin build or something obscure
like that.