nnrpd: New groupexactcount parameter to tune article counts
Added a new groupexactcount parameter in readers.conf to force nnrpd to
recount and report the exact number of still existing articles in newsgroups
considered to have an estimate number of articles below groupexactcount (set
to 5 by default).  The primary goal of this recount is to advertise newsgroups
as empty when they are (articles that have been cancelled or self-expired may
otherwise still be in counted in the estimate).

closes #228