nnrpd: Speed up searches when gaps in article numbers
LAST and NEXT, as well as (X)HDR and XPAT when the requested header field is
not in the overview, try to retrieve each article number, one by one, with
ARTopen.

On newsgroups with huge gaps in article numbers, like 1 million, these
commands take a long time to respond.  Several minutes in a newsgroup with 2
articles, numbered 1 and 2147483647 for instance.

We now use OVopensearch to directly retrieve the information of available
article numbers in overview data and then verify that the article exists with
ARTopen (like it is currently done).  The responses of the above NNTP commands
are then immediate!

close #249