ovsqlite: Fix latent bug in write_request()
data, and not request->data, is expected to be written.
They are the same except in a rare event: when the request is larger than
the socket's output buffer AND a signal arrives while the write call blocks.

Bug found with a "variable set but not used" gcc warning.

Fix confirmed by Bo Lindbergh, author of ovsqlite.