Add Auto-Submitted header field in outgoing mails
Added a new "-a" option to innmail to specify additional header fields
to add in the headers of messages.  This is notably used to internally
support the addition of the Auto-Submitted header field in outgoing
mails.

All the programs sending mails now include an Auto-Submitted header
field in the message headers (either set to "auto-generated" or
"auto-replied", following the recommendation in RFC 3834):

- Daily Usenet reports (news.daily);

- newgroup/rmgroup/checkgroups reports (controlchan);

- top1000 reports (sendinpaths);

- failure or warning reports from various programs (mailpost, expirerm,
rc.news, innwatch, writelog);

- our internal wrapper to sendmail or like (innmail);

- gatewaying (news2mail);

- forwarding to a moderator (nnrpd).

Thanks to Harald Dunkel for this suggestion which will for instance
help to avoid unnecessary vacation replies.