pullnews: Gracefully handle socket timeout
pullnews goes on downloading articles one by one from a remote news
server even though the connection was stopped for whatever reason.  It
now checks for that, and makes re-attempts according to the -t setting.

Net::NNTP returns a 421 (sic) pseudo response code when a connection is no
longer active; that's what pullnews checks for.

Thanks to Jesse Rehmer for the bug report.