Fix the use of Auto-Submitted header field
Use auto-generated for controlchan reports, and do not use that header
field at all for gatewaying (news2mail and nnrpd).

Thanks, Russ Allbery, for your wise thoughts:

The notifications sent by controlchan are not an autoreply to the
sender of the control message.  They are a status report sent to the
news administrator about a change made (or proposed) on their news
server by a third party.  As such, they're equivalent to the daily
report or the notification messages from innwatch.  The fact that the
change to the server was initiated by a netnews message doesn't feel
horribly relevant.  A "reply" by definition goes back to the sender
of the message; notification messages that go to third parties feel
inherently different.

Gatewaying (of which moderation is a subset) is yet a different thing,
and the Auto-Submitted header isn't appropriate here with any value.
These messages are not Auto-Submitted; they're, well, gatewayed.
This is equivalent to email forwarding; mail servers don't add
Auto-Submitted headers when they forward mails.

*If* we were encapsulating messages when sending them to moderators,
"Auto-Submitted: auto-generated" may be in the outer envelope, but
since we're transforming and forwarding a message, this is just part
of the message propagation path and thus should not be labeled with
Auto-Submitted.  The message is not, in general, auto-submitted; it was
manually posted by the poster of the message, and we're just passing it
along to its ultimate recipient.

Also, on a more practical level, addition of this header is
potentially surprising to moderators.  Their mail client may rightfully
downgrade or refile messages with Auto-Submitted headers, particularly
Auto-Submitted: auto-replied, since the vast majority of such messages
are vacation auto-replies that are of low importance or that people may
even auto-delete.  There's some risk of confusing submitted posts for a
moderated group, which are relatively important mail, for relatively
unimportant mail like vacation autoreplies or automated status
messages.

see #255