Check the syntax of the domain parameter
If domain (manually set, or the FQDN) has invalid characters in
inn.conf or readers.conf, just abort with a fatal error.  It is a basic
configuration error and we do not want to generate invalid Message-IDs.

Add a new IsValidDomain() function to validate the right-hand part of
a Message-ID.  Call it to validate the FQDN when domain is unset, and
the domain when set.

close #242