Keep user-supplied domain in readers.conf
Customizing the domain part of Message-IDs generated by nnrpd and the
server name indicated in Injection-Info header fields is now easier:
the domain parameter in the access blocks of readers.conf can be
directly used (without setting virtualhost as it was previously the
case).

An expicitly-set configuration option should override whatever is
discovered while probing around on the system.  So we now use domain
unconditionally if it is set in readers.conf for forming message
identifiers and Injection-Info.  But keep backwards-compatibility by
not using domain if only set in inn.conf (previous versions of INN only
use it if virtualhost is also set to true).

Document domain as a first-class configuration parameter for access
groups in readers.conf, saying that it changes the way Message-IDs and
Injection-Info are constructed.

Update wording in the FAQ and various documentation (inn.conf, libinn,
readers.conf, comments in C code).

close #243