Reorder and detail some release process actions
"make depend" was not done at the right time (it needs a build with all
--with-xxx flags to enable all headers).

"make reformat" should be done after "make depend" so as not to change
the order of headers of files generated by buildconfig.

Explain why version numbers should not be changed in FAQ too earlier.

Recall that config.guess, config.sub and install-sh are executables.