ovsqlite-util: Fix how transactions are handled
Turning enforcement of foreign key constraints on or off is an SQLite
operation that can only be done outside of any transaction.

When the AutoCommit attribute is false, the do method of the database
handle will helpfully start a transaction if one isn't active already.
That is why this AutoCommit attribute has to be set to true for that
operation.

Patch from Bo Lindbergh.