pullnews: Fix lock file handling
When pullnews is run outside INN as a separate script (rare use case),
the PID file was deleted when a second instance of pullnews was run in
parallel.  Just after reporting another instance was already running,
it deleted the PID file.

The behaviour is correct when using INN::Utils::Shlock inside an INN
installation.