Permit increasing FD_SETSIZE at build time
Add a -DLARGE_FD_SETSIZE option to set the value of FD_SETSIZE at
build time.  The default system value is sometimes not enough for busy
news servers.

Thanks to Jesse Rehmer for the bug report.  Patch taken from Diablo,
and adapted for INN.

The long-term solution is of course to switch to libevent, as this
patch is a terrible hack.

see #273