innd: Ensure channels do not have a large fd
innd no longer malfunctions nor throttles when the maximum number of
file descriptors supported by the system is reached.  It was happily
select'ing on those and corrupting data structures in memory.

Add a new function isvalidfd() in lib/fdlimit.c used by innfeed and
innd.  Call it before CHANcreate() and NCcreate() so that they do not
create a channel with a too high file descriptor.

Also limit the file descriptors to (sizeof(fd_set) * CHAR_BIT - 1) when
FD_SETSIZE is not set.

Thanks to Jesse Rehmer for the bug report.

close #273