Include portable/system.h as first portable header
In a block of included headers, sort them with config.h first, followed
with portable/system.h and then the other headers in alphabetical
order.

Otherwise some defines in portable/system.h which could be used by
other headers may not be defined.  It notably affects the forced
FD_SETSIZE redefinition which was not taken into account when
portable/socket.h was included before portable/system.h.  This is now
fixed.

Thanks to Jesse Rehmer for the bug report.

see #273