Refuse headers with a leading dot in OVERVIEW.FMT
Header field names beginning with a dot can no longer be added in
the extraoverviewadvertised and extraoverviewhidden parameters of
inn.conf so as to fix problems of searches and dot-stuffing (in LIST
OVERVIEW.FMT responses).

This is not (yet) supported in INN, and as this use case does not
happen in real life, do not bother supporting it.

see #286