ovsqlite-util: Fix database lock issue (timeout)
Raise the busy timeout value to the same as the one of ovsqlite-server.
The DBD::SQLite default is 30 seconds, and is not enough when the
transtimelimit setting in ovsqlite.conf is higher than this value.

Thanks to Jesse Rehmer for the bug report and Bo Lindbergh for the fix.

close #282