New INN::ovsqlite_client Perl module
INN::ovsqlite_client permits accessing an ovsqlite overview database
through ovsqlite-server from a Perl script.

This module provides search, add, remove and expire functions for
information stored in an ovsqlite database (newsgroups and overview
data associated to articles in these newsgroups).

Many thanks to Bo Lindbergh for it!  That's a great and useful work.

Two samples are also provided in the contrib directory (ovsqlite-dump
and ovsqlite-undump) to show how to use the module.