ovsqlite: Check the length of added overview data
At read-time (search), ovsqlite-server implements a limit of 100,000
bytes to detect corrupt overview data.  This patch adds the same limit
when adding overview data, so that we do not add overview data into the
database that could not be read afterwards.

Thanks to Jesse Rehmer for the bug report (a newsgroup couldn't be
expired because of a spam with 111,497 bytes of overview data that
had been added into the database but was considered as corrupt when
expiring the newsgroup).

close #293