Remove stripping IP options from innd manual
A bug about stripping IP options not working for IPv6 connections was
mentioned in the innd manual page.

Stripping IP options for IPv4 was a vestige introduced in INN 1.6b1
and already removed in INN 2.6.0 in 2014 when modernizing the network
library.

Anyway, even the initial code from INN 1.6b1 was incomplete and did
not prevent INN from being vulnerable to IP spoofing attacks with
source-routed TCP connections.

Other software that also does IP-based restrictions never bothers with
this: xinetd, for example, or the rsync server.  That implies that this
is no longer considered a significant security issue, probably because
it's been fixed upstream in the kernel network stack.

close #183
(more information in the ticket)