innreport: Fix the totals in readership stats
Elapsed times, read articles and groups, etc. were wrong in totals.

When innreport tags a reader as a curious one (that is to say a
connection which does not send NNTP commands for reading nor posting)
or finds out that the client did not manage to authenticate, it was
removed from the readership statistics by client but is still counted
in the statistics per domain.  Now fixed.

Also remove some unused variables while refactoring the related code.

close #306