ovsqlite-util: Remove useless check on a return
It doesn't seem to be explicitly stated in the current DBI
documentation, but the RaiseError attribute passed to the connect
method applies to that method call itself as well as to future method
calls on the returned database handle.  So, with RaiseError => 1,
there's no need to check the return value.

This behaviour has been present since DBI version 0.91 released in
1997.