Properly `apt-get install inn2` in a GitHub Action.
Demonstrate how to properly `apt-get install inn2` in a GitHub Action.
Configure the inn2 instance so that other code can test interactions with the inn2 server.

Much of this project's documentation covers how to build inn2 from source.  It would be helpful if there were working examples of how to `apt-get install inn2` in Docker or in a GitHub Action.  I have been struggling to do so for a few hours without success.