Do not bail when a FQDN is not found
Improve how the check works at startup:

- recognize an INN_HOSTNAME environment variable that, when set and
fully qualified, is taken by INN as the hostname;

- make non-fatal the check for a fully qualified hostname.  This way,
the warning appears in the logs, and the news admin will see it and
can take appropriate actions.  The installation of INN will no longer
fail when using the Debian package (and probably also from other
distributions) on Docker.

The second point is normally enough to solve the issue of this ticket.
The environment variable permits silencing the warning.

close #311