Move man pages of delayer, shlock, sm to section 8
Section 8 is better than section 1 for the manual pages of these
programs.

The rationale behind the change is to fix Debian bug 1101114 which
reported a conflict of files installed by the sm (screen-message)
package and by our inn2 (INN) package because both of them installed
sm(1).  We did not notice before as screen-message installed sm(6),
which was changed to sm(1) in a recent release.