perl-nocem: Fix the handling of continuation lines
The previous commit introduced a regression in the parsing of
continuation lines.  Hopefully it has not a real impact as current
NoCeM senders do not use that syntax, and even if they did, the
articles would still be correctly removed.  Only the very specific
local function called via the use of the -i option would not have the
whole list of newsgroups the article was posted to.