Drop old OpenSSL version support
The new baseline is OpenSSL 1.1.1 or LibreSSL 2.8.0. I chose the former
because it's the version available in Rocky Linux 8 (and presumably RHEL 8),
which is still under support. The latter is just because it's been released
around the same time.

By doing this, we can drop every preprocessor conditional that depends on
`OPENSSL_VERSION_NUMBER` and almost every one that depends on
`LIBRESSL_VERSION_NUMBER`.

In addition, we can now make the following simplifications:

* Drop support for SSLv2, since neither library supports it anymore.
  For backwards compatibility, the protocol name is still allowed in the
  config, but does nothing.

* Remove the setting of the `SSL_OP_SINGLE_{EC,}DH_USE` options, since
  they're now permanently enabled in both libraries.

* Remove the `OPENSSL_init_ssl` call, which is now unnecessary.

* Remove the manual seeding code, since both libraries now handle seeding
  automatically. In LibreSSL, `RAND_load_file` doesn't even do anything.