Improve multilib searches in some rare situations
Fix a FreeBSD PR reported upstream, and fixed in rra-c-util.

Always add -L<prefix>/lib in addition to -L<prefix>/lib{32,64} even if
the latter exists.  Then hopefully the correct multilib directory will
always be first in the search path and any 32-bit libraries in the
/lib directory will be ignored.  It may only increase the risk that
we incorrectly link with the 32-bit version of the library because
the Fedora-style multilib layout unfortunately used /lib for 32-bit
libraries on 64-bit systems.

close #326